+++ R/parse_output_arg.R
@@ -45,11 +45,15 @@ parse_output_arg <- function(output) {
   }
 
   # if knit to word
-  if (check_dependency("knitr") && check_dependency("rmarkdown")) {
-    fmt <- try(rmarkdown::default_output_format(knitr::current_input())$name, 
+  if (check_dependency("knitr") && 
+      check_dependency("rmarkdown")) {
+    fmt <- try(
+      rmarkdown::default_output_format(knitr::current_input())$name, 
       silent = TRUE)
     if (!inherits(fmt, "try-error")) {
-      if ("word_document" %in% fmt) {
+      word_fmt <- c("word_document", "rdocx_document",
+                    "officedown::rdocx_document")
+      if (any(word_fmt %in% fmt)) {
         output_format <- "word"
       }
     }