+++ NEWS.md
@@ -6,6 +6,7 @@
 * modelsummary_wide has a new "stacking" argument
 * gof_map accepts list of lists. only needs 3 columns.
 * support officedown Rmd
+* bug: passing arguments through ...
 * bug: stars and rounding
 
 # modelsummary 0.6.4